+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Features
 
 - Added support for filtering reports by client via `{INCLUDE|EXCLUDE} CLIENTS "Client A"[, "Client B", ...]`
+- Added support for grouping list reports by client via `GROUP BY CLIENT`
 
 ## [0.4.1]